using serve

+++ copy.js
@@ -14,19 +14,24 @@ const output = path.resolve(__dirname, 'media')
 
 const names = data.map(d => d.name)
 
-fs.removeSync(output)
-
 const files = fs.readdirSync(episodes)
     .filter(episode => names.includes(episode))
     .forEach(episode => {
         const dir = path.resolve(episodes, episode)
         const src = path.resolve(dir, `blankstring-${episode}.mp3`)
         const dest = path.resolve(output, `${episode}.mp3`)
-        fs.copySync(src, dest)
-        console.log(
-            chalk.yellow('copied'),
-            chalk.blueBright.underline(src),
-            chalk.yellow('to'),
-            chalk.blueBright.underline(dest)
-        )
+        if (!fs.existsSync(dest)) {
+            fs.copySync(src, dest)
+            console.log(
+                chalk.yellow('copied'),
+                chalk.blueBright.underline(src),
+                chalk.yellow('to'),
+                chalk.blueBright.underline(dest)
+            )
+        } else {
+            console.log(
+                chalk.yellow('skipping'),
+                chalk.blueBright.underline(dest)
+            )
+        }
     })
